Move variant info into its own file
