+++ client/editorCtrl.ts
@@ -4,13 +4,14 @@ import * as cg from 'chessgroundx/types';
 import * as util from 'chessgroundx/util';
 
 import { _ } from './i18n';
-import { validFen, hasCastling, unpromotedRole, promotedRole, notation } from './chess'
+import { validFen, hasCastling, unpromotedRole, promotedRole } from './chess'
 import { diff, calculatePieceNumber } from './material';
 import { iniPieces } from './pieces';
 import { copyBoardToPNG } from './png';
 import { patch } from './document';
 import { PyChessModel } from "./types";
 import { ChessgroundController } from './cgCtrl';
+import { notation } from './variants';
 
 export class EditorController extends ChessgroundController {
     model;
