+++ client/gameInfo.ts
@@ -1,11 +1,12 @@
 import { h, VNode } from "snabbdom";
 
 import { _ } from './i18n';
-import { colorIcon, VARIANTS } from './chess';
+import { colorIcon } from './chess';
 import { aiLevel, gameType, renderRdiff } from './result';
 import { timeago, } from './datetime';
 import { timeControlStr } from "./view";
 import { PyChessModel } from "./types";
+import { VARIANTS } from "./variants";
 
 
 export function gameInfo(model: PyChessModel): VNode {
