+++ client/sound.ts
@@ -3,7 +3,7 @@ import { h, VNode } from 'snabbdom';
 import { Howl } from 'howler';
 
 import { _ } from './i18n';
-import { Variant } from './chess';
+import { Variant } from './variants';
 import { StringSettings, NumberSettings } from './settings';
 import { radioList, slider } from './view';
