+++ client/tournament.ts
@@ -8,8 +8,8 @@ import { JSONObject, PyChessModel } from './types';
 import { _ } from './i18n';
 import { patch } from './document';
 import { chatMessage, chatView, IChatController } from './chat';
-//import { sound } from './sound';
-import { colorIcon, VARIANTS, uci2LastMove, Variant } from './chess';
+import { colorIcon, uci2LastMove } from './chess';
+import { VARIANTS, Variant } from './variants';
 import { timeControlStr } from "./view";
 import { initializeClock, localeOptions } from './tournamentClock';
 import { gameType } from './result';
